+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Current (in progress)
+## 3.0.4 (2021-08-12)
 
 - Now returning notfound() http exception when router does not match any territory object instead of None [#2637](https://github.com/opendatateam/udata/pull/2637)
 - Add larger reuse thumbnail image [#2638](https://github.com/opendatateam/udata/pull/2638)